Feat: 이벤트 생성, 이벤트 + 정기 모임 + 스케줄 조회
