+++ src/main/java/leets/weeth/domain/schedule/application/dto/ScheduleDTO.java
@@ -1,0 +1,14 @@
+package leets.weeth.domain.schedule.application.dto;
+
+import java.time.LocalDateTime;
+
+public class ScheduleDTO {
+
+    public record Response(
+            Long id,
+            String title,
+            LocalDateTime start,
+            LocalDateTime end,
+            Boolean isMeeting
+    ) {}
+}
