+++ src/main/java/leets/weeth/domain/schedule/application/usecase/MeetingUseCase.java
@@ -1,0 +1,9 @@
+package leets.weeth.domain.schedule.application.usecase;
+
+import static leets.weeth.domain.schedule.application.dto.MeetingDTO.Response;
+
+public interface MeetingUseCase {
+
+    Response find(Long eventId);
+
+}
