+++ src/main/java/leets/weeth/domain/schedule/domain/entity/Event.java
@@ -1,0 +1,19 @@
+package leets.weeth.domain.schedule.domain.entity;
+
+import jakarta.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@SuperBuilder
+public class Event extends Schedule {
+
+    private String requiredItem;
+
+}
