+++ src/main/java/leets/weeth/domain/schedule/domain/entity/Meeting.java
@@ -1,0 +1,23 @@
+package leets.weeth.domain.schedule.domain.entity;
+
+import jakarta.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@SuperBuilder
+public class Meeting extends Schedule {
+
+    private Integer weekNumber;
+
+    private Integer cardinal;
+
+    private Integer code;
+
+}
