+++ src/main/java/leets/weeth/domain/schedule/domain/repository/EventRepository.java
@@ -1,0 +1,12 @@
+package leets.weeth.domain.schedule.domain.repository;
+
+import leets.weeth.domain.schedule.domain.entity.Event;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface EventRepository extends JpaRepository<Event, Long> {
+
+    List<Event> findByStartLessThanEqualAndEndGreaterThanEqualOrderByStartAsc(LocalDateTime end, LocalDateTime start);
+}
