+++ src/main/java/leets/weeth/domain/schedule/domain/service/MeetingSaveService.java
@@ -1,0 +1,4 @@
+package leets.weeth.domain.schedule.domain.service;
+
+public class MeetingSaveService {
+}
