+++ src/main/java/leets/weeth/domain/user/application/converter/CardinalListConverter.java
@@ -1,0 +1,35 @@
+package leets.weeth.domain.user.application.converter;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.persistence.AttributeConverter;
+
+import java.io.IOException;
+import java.util.List;
+
+public class CardinalListConverter implements AttributeConverter<List<Integer>, String> {
+
+    private static final ObjectMapper mapper = new ObjectMapper()
+            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+            .configure(DeserializationFeature.FAIL_ON_NULL_FOR_PRIMITIVES, false);
+
+    @Override
+    public String convertToDatabaseColumn(List<Integer> attribute) {
+        try {
+            return mapper.writeValueAsString(attribute);
+        } catch (JsonProcessingException e) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    @Override
+    public List<Integer> convertToEntityAttribute(String dbData) {
+        try {
+            return mapper.readValue(dbData, new TypeReference<>() {});
+        } catch (IOException e) {
+            throw new IllegalArgumentException();
+        }
+    }
+}
