+++ src/main/java/leets/weeth/domain/user/domain/entity/enums/Position.java
@@ -1,0 +1,7 @@
+package leets.weeth.domain.user.domain.entity.enums;
+
+public enum Position {
+    D,
+    FE,
+    BE
+}
