+++ src/main/java/leets/weeth/domain/user/domain/entity/enums/Role.java
@@ -1,0 +1,11 @@
+package leets.weeth.domain.user.domain.entity.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+    USER,
+    ADMIN
+}
