+++ src/main/java/leets/weeth/domain/user/domain/entity/enums/Status.java
@@ -1,0 +1,8 @@
+package leets.weeth.domain.user.domain.entity.enums;
+
+public enum Status {
+    WAITING,
+    ACTIVE,
+    BANNED,
+    LEFT
+}
