--- src/main/java/leets/weeth/domain/user/entity/enums/Position.java
@@ -1,7 +1,0 @@
-package leets.weeth.domain.user.entity.enums;
-
-public enum Position {
-    D,
-    FE,
-    BE;
-}
