--- src/main/java/leets/weeth/domain/user/entity/enums/Role.java
@@ -1,11 +1,0 @@
-package leets.weeth.domain.user.entity.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-    USER,
-    ADMIN
-}
