--- src/main/java/leets/weeth/domain/user/entity/enums/Status.java
@@ -1,8 +1,0 @@
-package leets.weeth.domain.user.entity.enums;
-
-public enum Status {
-    WAITING,
-    ACTIVE,
-    BANNED,
-    LEFT
-}
