+++ src/main/java/leets/weeth/global/auth/login/filter/CustomJsonUsernamePasswordAuthenticationFilter.java
@@ -3,7 +3,7 @@ package leets.weeth.global.auth.login.filter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import leets.weeth.domain.user.repository.UserRepository;
+import leets.weeth.domain.user.domain.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationServiceException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
